requirements.txt
openpyxl>=3.1.0
pystray>=0.19.0
Pillow>=10.0.0
plyer>=2.1.0
pyinstaller>=6.0.0
